mytelegrambot/requirements.txt
aiogram~=2.21
python-telegram-bot==12.0.0b1
